move hooks back into components
