+++ packages/ui-components/src/components/ChartOverlayX/index.ts
@@ -1,1 +1,2 @@
 export * from "./ChartOverlayX";
+export * from "./useHoveredDate";
