+++ packages/ui-components/src/components/ChartOverlayXY/ChartOverlayXY.stories.tsx
@@ -5,8 +5,7 @@ import { scaleUtc, scaleLinear } from "@visx/scale";
 import { Timeseries } from "@actnowcoalition/metrics";
 import { AxisBottom, AxisLeft } from "../Axis";
 import { LineChart } from "../LineChart";
-import { useHoveredPoint } from "../../common/hooks";
-import { ChartOverlayXY } from ".";
+import { ChartOverlayXY, useHoveredPoint } from ".";
 
 export default {
   title: "Charts/ChartOverlayXY",
