+++ packages/ui-components/src/components/ChartOverlayXY/index.ts
@@ -1,1 +1,2 @@
 export * from "./ChartOverlayXY";
+export * from "./useHoveredPoint";
