+++ packages/ui-components/src/components/MetricLineThresholdChart/MetricLineThresholdChart.tsx
@@ -6,10 +6,10 @@ import max from "lodash/max";
 import { assert } from "@actnowcoalition/assert";
 import { Region } from "@actnowcoalition/regions";
 import { Metric } from "@actnowcoalition/metrics";
-import { useData, useHoveredDate } from "../../common/hooks";
+import { useData } from "../../common/hooks";
 import { AxesTimeseries } from "../AxesTimeseries";
 import { GridRows } from "../Grid";
-import { ChartOverlayX } from "../ChartOverlayX";
+import { ChartOverlayX, useHoveredDate } from "../ChartOverlayX";
 import { useMetricCatalog } from "../MetricCatalogContext";
 import { MetricTooltip } from "../MetricTooltip";
 import { BaseChartProps } from "../../common/utils/charts";
